fix(ZMS): move log display formatting out of zmsadmin zmsdb import
zmsadmin must not depend on zmsdb; formatDisplayFields now lives on the Log entity and zmsdb delegates to it.
